AS-3: Added required account service methods along with mapper and required database tables

+++ account-service/src/main/resources/db/migrations/V1__create_account.sql
@@ -1,5 +1,5 @@
 CREATE TABLE account(
-id bigint PRIMARY KEY,
+id serial PRIMARY KEY,
 account_number bigint NOT NULL,
 account_type varchar(20) NOT NULL,
 account_balance decimal(10,2) NOT NULL
