+++ account-service/src/main/resources/db/migrations/V2__create_branch.sql
@@ -1,5 +1,5 @@
 CREATE TABLE branch(
-id bigint PRIMARY KEY,
+id serial PRIMARY KEY,
 branch_code bigint NOT NULL,
 branch_name varchar(50) NOT NULL,
 branch_address varchar(255) NOT NULL,
